+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.5.0]
+### Changed
+- Reworked scripts code convention
+- CouchMultiplayerPlayer is now the base class for players
+
+### Removed
+- CouchMultiplayerPlayerUI
+- CouchMultiplayerPlayerUIManager
+
 ## [1.4.0]
 ### Added
 - CouchMultiplayerLobby.cs